% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-df.R
\name{print.data.frame}
\alias{print.data.frame}
\title{Printing Data Frames}
\usage{
\method{print}{data.frame}(df, n_extra = 20, n = 9, ...)
}
\arguments{
\item{df}{Data frame.}

\item{n_extra}{How many extra columns to print. These are the columns that
did not fit the console width, so only their names and type/class
are printed in the footer. Defaults to 20.}

\item{n}{How many rows to print. Defaults to 9.}

\item{...}{Additional arguments. Currently ignored, unless you revert back
to base R printing via options (see details).}
}
\value{
df, invisibly
}
\description{
This method overloads \code{base::print.data.frame()} to provide a nicer print
method for data frames.
}
\examples{
iris
}
\section{Options}{

You can disable selected features or revert back to base R method via options.
This is important because some editors may not support certain ANSI features
such as bold font. At the moment, the print method only uses 3 types of ANSI
formatting:
\itemize{
\item use \code{options(dfprint_strong = FALSE)} to disable bold font, used in variable
names
\item use \code{options(dfprint_italic = FALSE)} to disable italic, used for character
vector elements and column summaries
\item use \code{options(dfprint_subtle = FALSE)} to disable the grey font, used for
list columns, row names, missing and NULL values, factor orderings,
and column summaries
}

To revert back to base R print method, use \code{options(dfprint_enable = FALSE)}.
This may be useful if detaching the package is for some reason not desirable.
}
\section{How it prints}{

See \code{vignette("print-data-frames")}.
}
